Cloud Computing/PredictCommunity/requirements.txt
# Function dependencies, for example:
# package>=version
numpy==1.20.1
requests==2.25.1
tensorflow==2.4.0
google-cloud-storage==1.38.0
pandas==1.2.4
scikit_learn==0.24.2
keras==2.4.0